Changes in modal layout
