+++ package.json
@@ -14,6 +14,7 @@
     "@editorjs/header": "^2.8.1",
     "@editorjs/image": "^2.9.0",
     "@editorjs/list": "^1.9.0",
+    "@editorjs/marker": "^1.4.0",
     "@editorjs/paragraph": "^2.11.4",
     "@radix-ui/react-icons": "^1.3.0",
     "@remixicon/react": "^4.1.1",
