+++ src/components/layout/navbar/MobileNav.tsx
@@ -18,7 +18,7 @@ const MobileNav = () => {
 
   return (
     <>
-      <div className='sticky left-0 top-0 flex w-full items-center justify-between gap-5 bg-primary-monkeyWhite/75 px-5 py-2 backdrop-blur-md dark:bg-primary-monkeyBlack/75 z-30'>
+      <div className='sticky left-0 top-0 flex w-full items-center justify-between gap-5 bg-primary-monkeyWhite/75 px-5 py-2 backdrop-blur-sm dark:bg-primary-monkeyBlack/75 z-30'>
         <Link href='/'>
           <Logo showMobileLogo={true} />
         </Link>
@@ -35,7 +35,7 @@ const MobileNav = () => {
         </div>
       </div>
 
-      <div className='fixed bottom-0 left-0 flex w-full items-center justify-evenly bg-primary-monkeyWhite/50 px-5 py-4 backdrop-blur-sm dark:bg-primary-monkeyBlack/50 z-50'>
+      <div className='fixed bottom-0 left-0 flex w-full items-center justify-evenly bg-primary-monkeyWhite px-5 py-4 dark:bg-primary-monkeyBlack z-50 border-t-1 border-secondary-lightGrey/25'>
         <ThemeSwitch />
 
         <CreateButton showTitle={false} />
