+++ src/components/modals/layout/ModalHeader.tsx
@@ -1,31 +1,20 @@
 import { FC } from 'react';
 
-import IconContainer from '@/components/icon';
-
 type ModalHeaderProps = {
   heading?: string;
   subHeading?: string;
   showHeading?: boolean;
-  setModal: React.Dispatch<React.SetStateAction<boolean>>;
 };
 
 const ModalHeader: FC<ModalHeaderProps> = ({
   heading = 'Welcome to Monkeys',
   showHeading,
   subHeading,
-  setModal,
 }) => {
   return (
-    <div className='flex flex-col justify-center'>
-      <div className='m-2 flex justify-end'>
-        <IconContainer
-          name='RiCloseLine'
-          size={20}
-          onClick={() => setModal(false)}
-        />
-      </div>
+    <div className='mt-4 flex flex-col justify-center'>
       {showHeading && (
-        <div className='mt-4 flex flex-col gap-2'>
+        <div className='flex flex-col gap-2'>
           <p className='text-center font-playfair_Display text-2xl sm:text-3xl font-semibold'>
             {heading}
           </p>
