--- src/components/modals/login/loginSteps.ts
@@ -1,18 +1,0 @@
-import { LoginStep } from './LoginModal';
-
-export const loginSteps: LoginStep[] = [
-  {
-    id: 1,
-    heading: 'Account Login',
-    subHeading: 'choose an option to continue',
-  },
-  {
-    id: 2,
-    heading: 'Log in',
-    subHeading: 'to continue to Monkeys',
-  },
-  {
-    id: 3,
-    heading: 'Welcome',
-  },
-];
