--- src/components/modals/publish/publishSteps.ts
@@ -1,19 +1,0 @@
-import { PublishStep } from './PublishModal';
-
-export const publishSteps: PublishStep[] = [
-  {
-    id: 1,
-    heading: 'Blog Preview',
-    subHeading: 'add preview details',
-  },
-  {
-    id: 2,
-    heading: 'Blog Details',
-    subHeading: 'select topics suitable for your blog',
-  },
-  {
-    id: 3,
-    heading: '',
-    subHeading: '',
-  },
-];
