+++ src/components/modals/signup/Step2.tsx
@@ -2,10 +2,10 @@ import { FC, useState } from 'react';
 
 import Button from '@/components/button';
 import Input from '@/components/input';
+import { signupSteps } from '@/constants/modal';
 
 import ModalContent from '../layout/ModalContent';
 import { SignupStep } from './SignupModal';
-import { signupSteps } from './signupSteps';
 
 type Step2Props = {
   setLoginStep: React.Dispatch<React.SetStateAction<SignupStep>>;
