+++ src/components/modals/signup/Step3.tsx
@@ -3,10 +3,10 @@ import { FC, useState } from 'react';
 import Button from '@/components/button';
 import Icon from '@/components/icon/Icon';
 import Input from '@/components/input';
+import { signupSteps } from '@/constants/modal';
 
 import ModalContent from '../layout/ModalContent';
 import { SignupStep } from './SignupModal';
-import { signupSteps } from './signupSteps';
 
 type Step3Props = {
   setLoginStep: React.Dispatch<React.SetStateAction<SignupStep>>;
