--- src/components/modals/signup/signupSteps.ts
@@ -1,22 +1,0 @@
-export const signupSteps = [
-  {
-    id: 1,
-    heading: 'Join Monkeys',
-    subHeading: 'choose an option to continue',
-  },
-  {
-    id: 2,
-    heading: 'Basic Details',
-    subHeading: 'enter first name and last name',
-  },
-  {
-    id: 3,
-    heading: 'Login Details',
-    subHeading: 'set email address',
-  },
-  {
-    id: 4,
-    heading: 'Almost there',
-    subHeading: 'secure your account',
-  },
-];
