+++ src/config/editorjs.config.ts
@@ -3,11 +3,11 @@ import { EditorConfig } from '@editorjs/editorjs';
 const editorConfig: EditorConfig = {
   holder: 'editorjs-container',
   autofocus: true,
+  inlineToolbar: false,
   tools: {
     delimiter: require('@editorjs/delimiter'),
     header: {
       class: require('@editorjs/header'),
-      inlineToolbar: true,
       config: {
         placeholder: 'add a title...',
         levels: [1, 2, 3],
@@ -16,7 +16,6 @@ const editorConfig: EditorConfig = {
     },
     paragraph: {
       class: require('@editorjs/paragraph'),
-      inlineToolbar: true,
       config: {
         placeholder: 'add some text...',
       },
@@ -33,6 +32,10 @@ const editorConfig: EditorConfig = {
           byUrl: 'http://localhost:3000/api/fetchUrl',
         },
       },
+    },
+    Marker: {
+      class: require('@editorjs/marker'),
+      shortcut: 'CMD+SHIFT+M',
     },
   },
   defaultBlock: 'paragraph',
